Reading patient records from database instead of contacts
